+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 - anonymous without chat
 - offline viewer for anonymous
-- zh-TW, pt-PT
+- changed jwt implementation in csharp, csharp-mvc, php, ruby
+- added hy, eu, zh-TW, ms, pt-PT
 
 ## 1.2.0
 - ruby v3.0